Chapter 02 update

+++ 02_chapter/c02e07.c
@@ -1,13 +1,12 @@
 /* Denominations of United States currency */
 
 #include <stdio.h>
-#include <inttypes.h>
 
 int main(void)
 {
-    uint16_t amount;
+    unsigned int amount;
     printf("Enter a dollar amount: ");
-    scanf("%" SCNu16, &amount);
+    scanf("%u", &amount);
 
     printf("Nominal 20: %u\n", amount / 20);
     printf("Nominal 10: %u\n", (amount % 20) / 10);
